Move match to utils
